Basic Model for separation of leopards working.
Still a very raw commit, but the model is running seemingly successfully, the losses are decreasing. We try to use a very modular separation with very clear functions. train.py orchestrates everything, but trainer.py has the training loops. Backbone is a simple pretrained network which you can load from timm. I tested with resnet18 and resnet50 and it works well. Add an extra layer to embed to number of dimensions, and then in the TripletLoss.py, is where the magic happens. Here, we iterate through the elements of the batch, find all the positives, and a random negative for now, and calculate the triplet loss. With this we can backpropagate. Finally when the model has been trained, we run a visualization which attempts to plot the points according to their tags, running a t-SNE to decrease embedding dimension from 256 to 2 dimensions. Of course, we still have a long way to go for a full basic model to be working, we are missing evaluation metrics, and clustering to turn embeddings into new leopards.